+++ clients/gtest/sygvdx_hegvdx_gtest.cpp
@@ -129,22 +129,22 @@ class HEGVDX : public SYGVDX_HEGVDX<API_NORMAL>
 
 // non-batch tests
 
-TEST_P(SYGVDX, __float)
+TEST_P(SYGVDX, DISABLED__float)
 {
     run_tests<false, false, float>();
 }
 
-TEST_P(SYGVDX, __double)
+TEST_P(SYGVDX, DISABLED__double)
 {
     run_tests<false, false, double>();
 }
 
-TEST_P(HEGVDX, __float_complex)
+TEST_P(HEGVDX, DISABLED__float_complex)
 {
     run_tests<false, false, rocblas_float_complex>();
 }
 
-TEST_P(HEGVDX, __double_complex)
+TEST_P(HEGVDX, DISABLED__double_complex)
 {
     run_tests<false, false, rocblas_double_complex>();
 }