+++ library/src/amd_detail/dlopen/rocsparse.hpp
@@ -27,6 +27,12 @@
 #include <rocsparse/rocsparse.h>
 #else
 
+#if defined(rocsparse_ILP64)
+typedef int64_t rocsparse_int;
+#else
+typedef int32_t rocsparse_int;
+#endif
+
 // type definitions
 typedef enum rocsparse_status_
 {
@@ -61,11 +67,30 @@ typedef enum rocsparse_matrix_type_
     rocsparse_matrix_type_triangular = 3 /**< triangular matrix type. */
 } rocsparse_matrix_type;
 
+typedef struct _rocsparse_handle*    rocsparse_handle;
 typedef struct _rocsparse_mat_descr* rocsparse_mat_descr;
+
+typedef struct
+{
+    float x, y;
+} rocsparse_float_complex;
+
+typedef struct
+{
+    double x, y;
+} rocsparse_double_complex;
 
 HIPSOLVER_BEGIN_NAMESPACE
 
 // function declarations
+typedef rocsparse_status (*fp_rocsparse_create_handle)(rocsparse_handle* handle);
+extern fp_rocsparse_create_handle g_rocsparse_create_handle;
+#define rocsparse_create_handle ::hipsolver::g_rocsparse_create_handle
+
+typedef rocsparse_status (*fp_rocsparse_destroy_handle)(rocsparse_handle handle);
+extern fp_rocsparse_destroy_handle g_rocsparse_destroy_handle;
+#define rocsparse_destroy_handle ::hipsolver::g_rocsparse_destroy_handle
+
 typedef rocsparse_status (*fp_rocsparse_create_mat_descr)(rocsparse_mat_descr* descr);
 extern fp_rocsparse_create_mat_descr g_rocsparse_create_mat_descr;
 #define rocsparse_create_mat_descr ::hipsolver::g_rocsparse_create_mat_descr
@@ -81,6 +106,54 @@ extern fp_rocsparse_get_mat_type g_rocsparse_get_mat_type;
 typedef rocsparse_index_base (*fp_rocsparse_get_mat_index_base)(rocsparse_mat_descr descr);
 extern fp_rocsparse_get_mat_index_base g_rocsparse_get_mat_index_base;
 #define rocsparse_get_mat_index_base ::hipsolver::g_rocsparse_get_mat_index_base
+
+typedef rocsparse_status (*fp_rocsparse_scsr2dense)(rocsparse_handle          handle,
+                                                    rocsparse_int             m,
+                                                    rocsparse_int             n,
+                                                    const rocsparse_mat_descr descr,
+                                                    const float*              csr_val,
+                                                    const rocsparse_int*      csr_row_ptr,
+                                                    const rocsparse_int*      csr_col_ind,
+                                                    float*                    A,
+                                                    rocsparse_int             ld);
+extern fp_rocsparse_scsr2dense g_rocsparse_scsr2dense;
+#define rocsparse_scsr2dense ::hipsolver::g_rocsparse_scsr2dense
+
+typedef rocsparse_status (*fp_rocsparse_dcsr2dense)(rocsparse_handle          handle,
+                                                    rocsparse_int             m,
+                                                    rocsparse_int             n,
+                                                    const rocsparse_mat_descr descr,
+                                                    const double*             csr_val,
+                                                    const rocsparse_int*      csr_row_ptr,
+                                                    const rocsparse_int*      csr_col_ind,
+                                                    double*                   A,
+                                                    rocsparse_int             ld);
+extern fp_rocsparse_dcsr2dense g_rocsparse_dcsr2dense;
+#define rocsparse_dcsr2dense ::hipsolver::g_rocsparse_dcsr2dense
+
+typedef rocsparse_status (*fp_rocsparse_ccsr2dense)(rocsparse_handle               handle,
+                                                    rocsparse_int                  m,
+                                                    rocsparse_int                  n,
+                                                    const rocsparse_mat_descr      descr,
+                                                    const rocsparse_float_complex* csr_val,
+                                                    const rocsparse_int*           csr_row_ptr,
+                                                    const rocsparse_int*           csr_col_ind,
+                                                    rocsparse_float_complex*       A,
+                                                    rocsparse_int                  ld);
+extern fp_rocsparse_ccsr2dense g_rocsparse_ccsr2dense;
+#define rocsparse_ccsr2dense ::hipsolver::g_rocsparse_ccsr2dense
+
+typedef rocsparse_status (*fp_rocsparse_zcsr2dense)(rocsparse_handle                handle,
+                                                    rocsparse_int                   m,
+                                                    rocsparse_int                   n,
+                                                    const rocsparse_mat_descr       descr,
+                                                    const rocsparse_double_complex* csr_val,
+                                                    const rocsparse_int*            csr_row_ptr,
+                                                    const rocsparse_int*            csr_col_ind,
+                                                    rocsparse_double_complex*       A,
+                                                    rocsparse_int                   ld);
+extern fp_rocsparse_zcsr2dense g_rocsparse_zcsr2dense;
+#define rocsparse_zcsr2dense ::hipsolver::g_rocsparse_zcsr2dense
 
 HIPSOLVER_END_NAMESPACE